use size_t instead of unsigned int in one place
git-svn-id: https://svn.parrot.org/parrot/trunk@38482 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/debug.c
@@ -2756,7 +2756,7 @@ PDB_disassemble_op(PARROT_INTERP, ARGOUT(char *dest), size_t space,
         static const char regs[] = "ISPN";
 
         for (j = 0; j < n_values; j++) {
-            unsigned int idx = 0;
+            size_t idx = 0;
             const int sig_value = VTABLE_get_integer_keyed_int(interp, sig, j);
 
             /* Print the register name, e.g. P37. */
